Guard against blocks with just one terminator in HasSingleGraph()
We need to make sure the basic block has more than one ops before
checking whether the second op is a known terminator. Otherwise,
it will trigger an ASAN failure.

Also added a test for testing the degenerated case for both
executor to control and control to executor conversion, which
should cover the above case.

PiperOrigin-RevId: 260728798