Attempt to load versioned shared objects in Java
Since TensorFlow 1.14.0, libtensorflow_framework.so has been packaged as
a versioned shared object. This change attempts to support that naming
scheme in the Java NativeLibrary loader.

If libtensorflow_framework.so is not present and a major version number
can be determined, the loader will attempt to find
libtensorflow_framework.so.<majorVersion>, where <majorVersion> is
determined from the manifest of the JAR from which the NativeLibrary
class was loaded. On darwin, the loader will look for
libtensorflow_framework.<majorVersion>.so.

PiperOrigin-RevId: 260749010