[XLA] Rename Executable::SizeInBytes to Executable::SizeOfGeneratedCodeInBytes
On the GPU backend a lot of library code is used: cuDNN, cuBLAS, etc.
We do not know the size of the used kernels, so we can not report the total
amount of loaded kernels.

Renaming to SizeOfGeneratedCodeInBytes makes it clear that the returned number
does not include library code.

PiperOrigin-RevId: 261026023