Set narrow_range to true to quantize weights
This CL sets narrow_range to true to avoid the value -128 in in8 quantization,
thus the weight values range only in [-127, 127]. This enables faster runtime
arithmetic kernels on ARM NEON. For uint8 quantization, 128 is substracted from
the quantized values and zero-points and then the int8 kernels can be used,
thus the its narrow_range for weights are set to true as well. Note that the
FakeQuant* for "weights" inserted in all the existing models have narrow_range
set to true, so this CL just make it consistent for all the weights in the
model.

TOCO implemented the same logic in the
ensure_uint8_weights_safe_for_fast_int8_kernels pass. This optimizaion is very
specific to ARM architecture, so an TODO is added to make it configurable.

Activations shouldn't use narrow_range, instead, it should use the full range.

PiperOrigin-RevId: 261269551