[XLA] Better handling of tuples in memory space assignment.
Previously, tuples were not being handled properly. First, the tuple itself
(which is very cheap since it just has pointers) was being asynchronously
copied. For now, disable the tuple types from memory space assignment. In the
future, we should be able to place them both in default and alternate memory
spaces because they are cheap. Secondly, GetTupleElement instruction that
CopyStart depends on could have been scheduled after the CopyStart. This is
becuase the defining instruction of a tuple element buffer is the original
tuple, not the GetTupleElement instruction. Now, when we schedule a CopyStart,
we ensure its dependcies are scheduled before it.

PiperOrigin-RevId: 262027013