Rollback an accidental change that sneaked into commit 8cc20b.
Outcome of that change: for tf.cond(pred, lambda: outside_compilation(A), lambda: outside_compilation(B)), the "Switch" node for input of A will have outside compilation attribute, and we will then mark the whole functionalize If node with outside compialtion attribute. This effectively turns original code into outside_compilation(tf.cond(pred, A, B)), which causes all inputs of A/B to be transferred from TPU to host unconditionally.

We should still determine if the whole tf.cond() is in outside compilation by checking whether the predicate is in outside compilation.

PiperOrigin-RevId: 262424903