In UpdateTFE_ContextWithServerDef(), call GrpcServer::Start() after all other initialization.
Currently, GrpcServer does not support clean shutdown and destruction after GrpcServer::Start() is called, and logs a FATAL error in this case. In UpdateTFE_ContextWithServerDef(), there are several cases where a method can fail (e.g. due to invalid user input), and we attempt to destroy the created GrpcServer before returning. By deferring the call to GrpcServer::Start(), we can return an error to the client without crashing their process.

PiperOrigin-RevId: 262638388